Remove channel overrides; use Conda env channels
Prior to this commit, some API functions overrode the Conda channels to
explicitly include the `microdrop-plugins` channel.

As of this commit, all channel overrides are removed.  Instead use
channels configured from Conda environment only.  This provides the
greatest flexibility for sources of MicroDrop plugins and related
packages.  For example, a local package repository could be configured as
a channel for testing purposes.